by botraikova

+++ src/trianglesArea.java
@@ -2,8 +2,8 @@ public class trianglesArea {
 
     public static void main(String[] args) {
 
-        int height = 3;
-        int base = 4;
+        double height = 3;
+        double base = 4;
 
         double area = (height * base) / 2;
         System.out.println(area);
